requirements.txt
flask
sqlalchemy>=2.0
psycopg[binary]
gunicorn
python-dotenv
flask-login
werkzeug